Reworked update method added polys for player/mob
Polygons should now be around the player and the mob, and when
intersected with each other, the player will 'lose health'. rapidly.

Changed the update methods for player/mob to call one super method,
which will be found in their respective classes.

Messed with collision to see why up collision is so buggy. Still no fix.